scripts/forex_tester_importer/requirements.txt
# scripts/forex_tester_importer/requirements.txt
python-dotenv
boto3
pandas
pyarrow  # pandas.to_parquet のため